- eliminate unnecessary duplicated class - rename AbstractBindingItemVHFactory to BaseBindingItemVHFactory
