--- library-extensions-binding/src/main/java/com/mikepenz/fastadapter/binding/ModelAbstractBindingItem.kt
@@ -1,11 +1,0 @@
-package com.mikepenz.fastadapter.binding
-
-import androidx.viewbinding.ViewBinding
-import com.mikepenz.fastadapter.IModelItem
-
-/**
- * Implements an abstract model item simplifying usage with ViewBinding.
- */
-abstract class ModelAbstractBindingItem<VB : ViewBinding, Model, VH : BindingViewHolder<VB>> :
-        AbstractBindingItem<VB, VH>(), IModelItem<Model, VH>
-
